refactor: remove InputTagEditor and TagsDisplay components to simplify tag management
