+++ src/components/SinglePost/Tags/components/index.tsx
@@ -1,15 +1,11 @@
 import { InputTag } from './_InputTag';
 import { InputTagStandard } from './_InputTagStandard';
-import { InputTagEditor } from './_InputTagEditor';
-import { TagsDisplay } from './_TagsDisplay';
 import { InputTagSimple } from './_InputTagSimple';
 import { InlineTagInput } from './_InputTagInline';
 
 export const Components = {
   InputTag,
   InputTagStandard,
-  InputTagEditor,
-  TagsDisplay,
   InputTagSimple,
   InlineTagInput
 };
